% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_concept_for_denovo.R
\name{flag_concept_for_denovo}
\alias{flag_concept_for_denovo}
\title{Flags a concept for having a new concept created}
\usage{
flag_concept_for_denovo(manifest_id, concept_field, concept_name,
  reason = "")
}
\arguments{
\item{manifest_id}{identifier for concept}

\item{concept_field}{can only be "VARIABLE_FIELD_NAME" or "PERMISSIBLE_VALUE_LABEL"}

\item{concept_name}{concept as a character string of length 1}

\item{reason}{reason why you think denovo is needed}
}
\description{
Flags a concept for having a new concept created
}
