% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_concept_for_followup.R
\name{flag_concept_for_followup}
\alias{flag_concept_for_followup}
\title{Flags a concept for requiring further inquiry}
\usage{
flag_concept_for_followup(manifest_id, concept_field, concept_name,
  comment = "")
}
\arguments{
\item{manifest_id}{identifier for concept}

\item{concept_field}{can only be "VARIABLE_FIELD_NAME" or "PERMISSIBLE_VALUE_LABEL"}

\item{concept_name}{concept as a character string of length 1}

\item{comment}{your question about that concept that you need answered}
}
\description{
Flags a concept for requiring further inquiry
}
