% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_local_library.R
\name{query_local_library}
\alias{query_local_library}
\title{To reuse the same CUIs, first query the local library, which is a combination of an existing lookup found in data/ and the keys that are being generated in the working directory}
\usage{
query_local_library(...)
}
\arguments{
\item{...}{phrases used to query the local library}
}
\value{
dataframe of CUI and STR combinations from the lookup table and existing keys in the working directory
}
\description{
To reuse the same CUIs, first query the local library, which is a combination of an existing lookup found in data/ and the keys that are being generated in the working directory
}
