% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_multiple_cuis_for_one_concept.R
\name{save_multiple_cuis_for_one_concept}
\alias{save_multiple_cuis_for_one_concept}
\title{Saves multiple cuis to one concept into a KEY for pre-coordinated concepts}
\usage{
save_multiple_cuis_for_one_concept(manifest_id, concept_field,
  concept_name, ...)
}
\arguments{
\item{concept_field}{can only be "VARIABLE_FIELD_NAME" or "PERMISSIBLE_VALUE_LABEL"}

\item{...}{Multiple CUI-STRING combined in a vector. For example: c("C13431", "Breast cancer"), c("C39134", "Testicular cancer")}
}
\value{
dataframe written to a csv file prefixed with "Odysseus_KEY_REDCAP_TO_UMLS_COORDINATES_"
}
\description{
Saves multiple cuis to one concept into a KEY for pre-coordinated concepts
}
