% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_one_cui_that_matches_concept.R
\name{save_one_cui_that_matches_concept}
\alias{save_one_cui_that_matches_concept}
\title{Saves a cui to key}
\usage{
save_one_cui_that_matches_concept(manifest_id, concept_field, concept_name,
  cui, str)
}
\arguments{
\item{manifest_id}{identifier for concept}

\item{concept_field}{can only be "VARIABLE_FIELD_NAME" or "PERMISSIBLE_VALUE_LABEL"}

\item{concept_name}{concept as a character string of length 1}

\item{cui}{umls cui as a character string of length 1}

\item{str}{closest matching umls str as a characters string of length 1}
}
\description{
Saves a cui to key
}
